Refactor dashboard to use visitors-table component and update state management for events
